feat: добавила debounce
